Script cleanup

+++ server/src/handlers/cleanup/cleanup_database.ts
@@ -6,18 +6,27 @@ import { existsSync } from "fs";
 const CleanupDatabase = async (req: Request, res: Response): Promise<void> => {
   console.log("cleaning database from missing videos");
   let video_repo = getRepository(VideoMeta);
-  let videos = await video_repo.find();
+  let videos = await video_repo.find({ relations: ["scripts"] });
   let result = new Map<string, boolean>();
+  const counter = { deleted: 0, kept: 0, scripted: 0 };
   for (let v of videos) {
+    if (v.scripts && v.scripts.length > 0) {
+      console.log(`${v.id} has a script. Not deleting.`);
+      counter.scripted += 1;
+      continue;
+    }
     console.log(`checking videos ${v.id}`);
     let path_exists = existsSync(v.path);
     result.set(v.name, path_exists);
     if (!path_exists) {
       console.log(`deleting video ${v.id}`);
       await video_repo.remove(v);
+      counter.deleted += 1;
+    } else {
+      counter.kept += 1;
     }
   }
-  console.log("done cleanup missing videos");
+  console.log("done cleanup missing videos:", counter);
   res.status(200).json(Object.fromEntries(result));
 };
